+++ backend/Infra/Repositorio/PedidoRepository.cs
@@ -55,7 +55,7 @@ namespace Infra.Repositorio
                 {
                     ItemId = i.ItemId,
                     ItemNome = i.ItemNome,
-                    Categoria = (int)i.Categoria,
+                    Categoria = i.Categoria,
                     PrecoUnitario = i.PrecoUnitario,
                     Quantidade = i.Quantidade
                 }).ToList(),
